backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
langchain==0.0.312
google-generativeai
sqlalchemy==2.0.21
alembic==1.12.0
asyncpg==0.28.0
python-jose==3.3.0
pinecone-client==2.2.4
tiktoken==0.5.1
httpx==0.24.1
pytest==7.4.2